requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1

# Pydantic AI (for Chatbot Orchestration)
pydantic-ai==0.0.14

# OpenAI (for LLM in Chatbot Orchestration)
openai==1.54.3

# Google Gemini (for FileSearch)
google-generativeai==0.8.3
google-auth==2.31.0

# Web Scraping
crawl4ai==0.4.36
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
pydantic-settings==2.5.2
aiofiles==24.1.0
